Drafted a users configuration.
Perhaps we should consider not keeping all of these configs on the salt file server, but rather in
a corporate git; a shame we haven't one yet.

The current way of doing things does not allow for any flexibility, all users are created equal
and I'm not even sure I know how to make it otherwise.